% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_type.R
\name{assign.type}
\alias{assign.type}
\title{Assign spectra type using a logfile}
\usage{
assign.type(dn, logfile = "fieldlog.csv")
}
\arguments{
\item{dn}{A \pckge{hyperSpec} object}

\item{logfile}{Path to the corresponding logfile for the object}
}
\value{
A hyperSpec object including with sample identification [id] and indication if a reference measurement [is.REF]
}
\description{
Given a log file with columns sample identification, path to reference and sample measurements [id, R, S] this function creates a logical vector [is.REF] with the \pckge{hyperSpec}
}
\details{
The format of the logfile must be followed (.CSV). FIXME make more flexible
}
\examples{
dn <- scan.txt.SpectraSuite(files="data_for_tst/*.txt")
dn <- assign.type(dn, logfile="data_for_tst/fieldlog.csv")
}

