% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angle.R
\name{calcSolarZenith}
\alias{calcSolarZenith}
\title{Calculate solar zenith angle from timestamp}
\usage{
calcSolarZenith(time_stamp, latitude, longitude, timezone = 0)
}
\arguments{
\item{time_stamp}{A timestamp that can be parsed by POSIXct, such ISO8601. This must be UTC.}

\item{latitude}{Latitude of observer in decimal degrees, west is negative.}

\item{longitude}{Latitude of observer in decimal degrees, south is negative.}

\item{timezone}{Timezone of timestamp, west is negative.}
}
\value{
The Solar Zenith Angle in degrees
}
\description{
Given the timestamp, location and time zone calculate the solar zenith angle
for a observer. Requires package insol.
}
\examples{
time_stamp = "2014-11-26 11:30:26"
latitude = 36.529688
longitude = -6.292657
calcSolarZenith(time_stamp, latitude, longitude, timezone=0)
compare to https://www.nrel.gov/midc/solpos/spa.html
retrieved 2016-07-16: 11/26/2014,11:30:00,58.339855
}

