% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_area.R
\name{collection_area}
\alias{collection_area}
\title{Calculate surface area of the collection apature}
\usage{
collection_area(diameter, unit_in = "um")
}
\arguments{
\item{diameter}{The ‘diameter’ (d) of the instrument light collection
apature}

\item{unit_in}{The units of length for diameter, c('um', 'mm', 'cm', 'm')}
}
\value{
Surface area of the collection apature (*A*, m2)
}
\description{
Calculates the surface area using the diameter of the instrument light
collection apature. This maybe given in um, mm, cm or m and specified using
\code{unit_in}
}
\examples{
collection_area(400, unit_in='um')
}

