% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FASTSpectra.R
\docType{package}
\name{FASTspectra}
\alias{FASTspectra}
\alias{FASTspectra-package}
\title{Parse Ocean Optics spectra files.}
\description{
FASTSpectra allows you to parse multiple OO spectra
and include the instrument generated metadata. So far only "text with header" format is supported.
}
\details{
The only function you're likely to need from \pkg{FASTSpectra} is
\code{\link{scan.txt.SpectraSuite}}.
}

