% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{getSeason}
\alias{getSeason}
\title{Asign astronomical seasons}
\usage{
getSeason(DATES)
}
\arguments{
\item{DATES}{The dates to be binned into seasons}
}
\value{
A character vector of seasons
}
\description{
Asign astronomical seasons
}
\examples{
getSeason(c("2012-1-15", "2009-4-1", "2015-7-21", "2011-10-21"))
}

