% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperSpecToHsdar.R
\name{hyperSpecToHsdar}
\alias{hyperSpecToHsdar}
\title{Convert hyperSpec to hsdar object}
\usage{
hyperSpecToHsdar(spc)
}
\arguments{
\item{spc}{A hyperSpec object}
}
\value{
A hsdar SpecLib object
}
\description{
Convert hyperSpec to hsdar object
}
\examples{
# spc <- read.txt.wide.FAST(file = "FAST_ES_2_HCRF_2015-2016.txt", metadata = "FAST_ES_2_HCRF-metadata_2015-2016.txt")
#newspc <- hyperSpecToHsdar(spc)
}

