% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_IrradCal_uJCount.R
\name{import_IrradCal_uJCount}
\alias{import_IrradCal_uJCount}
\title{Import an OO 'IrradCal' instrument-specific response function}
\usage{
import_IrradCal_uJCount(file_path, theta_v = 0)
}
\arguments{
\item{file_path}{File path to a OO 'IrradCal' instrument-specific response function}

\item{theta_v}{The viewing geometry of measurements (θv, °),
expressed as an angle from the nadir (0°) viewing zenith.}
}
\value{
A \pkg{hyperSpec} object with the instrument-specific response 
function in units J / count and metadata for further radiometric conversions
}
\description{
Given the path to a specifically formatted IrradCal file in units of uJ /
count, parse the file extracts, tidies and calculates metadata, converts to
J and returns a \pkg{hyperSpec} object
}
\examples{
sys_path <- system.file("extdata", package = "FASTSpectra")
file_path <- Sys.glob(paste0(sys_path, "/*.IrradCal"))
cal <- import_IrradCal_uJCount(file_path)
summary(cal)
}

