% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_RefCal.R
\name{import_RefCal}
\alias{import_RefCal}
\title{Import a reference panel-specific reflectance response function}
\usage{
import_RefCal(file_path)
}
\arguments{
\item{file_path}{The file path of a reference panel-specific reflectance
response function}
}
\value{
A \pkg{hyperSpec} object containing the calibration data
}
\description{
Import a reference panel-specific reflectance response function
}
\examples{
sys_path <- system.file("extdata", package = "FASTSpectra")
file_path <- Sys.glob(paste0(sys_path, "/*.ReflCal"))
cal <- import_RefCal(file_path)
summary(cal)
}

