% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_irf.R
\name{normalise_irf}
\alias{normalise_irf}
\title{Calculate normalised instrument-specific response function}
\usage{
normalise_irf(irf, int_time, wavelengths)
}
\arguments{
\item{irf}{instrument-specific response function (IRFλ), usually with units J/count}

\item{int_time}{The ‘integration time’ (*t*) of a measurement i.e., the time period over which the instrument records light through the measuring apature. Usually seconds.}

\item{wavelengths}{Center points for the bands that make up the spectral measurement (λ) i.e., the wavelength scale. Usually units of nm.}
}
\value{
The normalised IRFλ vector.
}
\description{
Given an instrument-specific response function (IRFλ) in units Mass 
 squared Length per squared Time count, the 'integration time' of the calibration
 measurement in units Length and a vector of center points for the bands that
 make up the spectral measurement in units of Length this function will 
 return IRFnorm,λ in units of Mass cubed Length per Time count.  
 
 The usual case is IRFλ given in J/count
 
 Calculated as:
 \deqn{{IRF}_{\mathrm{norm, \lambda}} = \frac{t \times d_{\lambda}}{{IRF}_{\lambda}}}{IRFnorm,λ = t × dλ / IRFλ}
}
\examples{
normalise_irf(1:10, 0.001, 1:10)
}

