% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_collection_area.R
\name{proj_collection_area}
\alias{proj_collection_area}
\title{Calculate projected surface area of the collection apature}
\usage{
proj_collection_area(diameter, unit_in = "um", theta_v = 0)
}
\arguments{
\item{diameter}{The ‘diameter’ (d) of the instrument light collection
apature}

\item{unit_in}{The units of length for diameter, c('um', 'mm', 'cm', 'm')}

\item{theta_v}{The viewing geometry of measurements (θv, °),
expressed as an angle from the nadir (0°) viewing zenith in degrees.}
}
\value{
Projected area of the collection apature (*A*cosθ, m2)
}
\description{
Calculates the projected area using the diameter of the instrument light
collection apature and the viewing geometry of measurements (θv, °),
expressed as an angle from the nadir (0°) viewing zenith. Diameter maybe
given in um, mm, cm or m and specified using \code{unit_in}.
}
\examples{
proj_collection_area(400, unit_in='um', theta_v=0)
}

