% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_measurement_metadata.R
\name{QC_metatdata}
\alias{QC_metatdata}
\alias{chk_corr_elec_dark}
\alias{chk_corr_non_lin}
\title{Quality checking of measurement metadata}
\usage{
QC_metatdata(spc)

chk_corr_elec_dark(spc)

chk_corr_non_lin(spc)
}
\arguments{
\item{spc}{A hyperSpec object with specific metadata. See details}
}
\value{
Prints to console spectra row number that do not pass test or a 
message that all spectra passed.
}
\description{
Check that 'electrical dark' and 'non-linearity' corrections were applied
during a ground-based spectrometry measurement with an Ocean Optics 
spectrometer by accessing metadata flags.
}
\section{Functions}{
\itemize{
\item \code{chk_corr_elec_dark}: Check for electrical dark correction

\item \code{chk_corr_non_lin}: Check for non-linearity correction
}}
\examples{
#QC_metadata(spc) 
}

