% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_txt_wide_FAST.R
\name{read.txt.wide.FAST}
\alias{read.txt.wide.FAST}
\title{Parse FAST wide table ground-based spectrometry files}
\usage{
read.txt.wide.FAST(file, type = "reflectance", metadata = NULL)
}
\arguments{
\item{file}{Path to the spectral data. Wide table format with columns [id, timestamp, spc] and wavelength as spc column names.}

\item{type}{Reflectance, Irradiance or Radiance units}

\item{data}{Path to extra data, must have same number of rows as number of spectra}
}
\value{
A \pckg{hyperSpec} object
}
\description{
Parse FAST wide table ground-based spectrometry files
}
\examples{
# set path to data files
file.path <- system.file("extdata", package = "FASTSpectra")

# parse spectra into hyperSpec object
dn <- read.txt.OceanOptics(files=paste0(file.path,"/*.txt"))

# assign measurement id and type using fieldlog file
logfile <- paste0(file.path,"/fieldlog.csv")
dn <- assign.type(dn, logfile=logfile)

# split into sample and reference
type <- slot(dn,"data")[["type"]]
dn <- split(x=dn, f=type)

# convert samples to spectral radiance [W / sr m2 nm]
cal.rad <- paste0(file.path,"/*.IrradCal")
rad <- rad.corr(dn$SAMP, type="spectral.radiance", cal.DN2RadiantEnergy = cal.rad)
require(hyperSpec)
plot(rad, wl.range=380:850)

# convert references to spectral irradiance relative to 100\% reference panel [W / m2 nm]
cal.ref <- paste0(file.path,"/*.ReflCal")
ref <- rad.corr(dn$REF, type="spectral.irradiance", is.REF=TRUE, cal.DN2RadiantEnergy = cal.rad, cal.RRefPanel = cal.ref)
plot(ref, wl.range=380:850)

# calculate reflectance as 'radiance/irradiance'
SHR <- calc.reflectance(reference=ref, sample=rad, logfile=logfile)
plot(SHR, wl.range=380:850)

# write to file
# write.txt.wide.FAST(SHR, paste0(file.path,"/HCRF.spc"), paste0(file.path,"/HCRF-metadata.spc"))

# read txt.wide file
# read.txt.wide.FAST <- function(paste0(file.path,"/HCRF.spc"), type="reflectance", metadata=paste0(file.path,"/HCRF-metadata.spc"))
}

