% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_SpectraSuite_txt_2_hyperSpec.R
\name{scan.txt.SpectraSuite}
\alias{scan.txt.SpectraSuite}
\title{Parse Ocean Optics SpectraSuite "text with header"}
\usage{
scan.txt.SpectraSuite(files = "*.txt", label = list(spc = "DN (counts)"),
  ...)
}
\arguments{
\item{files}{A specific file path or list of file paths. Default is select all text files in working directory \code{"*.txt"}}

\item{label}{A list of form \code{list(parameter=expression(parameter))}. Default is to label the parameter as "DN (counts)"}
}
\value{
A hyperSpec object including a matrix of spectra, metadata extracted from the spectra headers and file information
}
\description{
Parse Ocean Optics SpectraSuite "text with header"
}
\examples{
setwd("~/Desktop/FASTSpectra") # DELETE ME
out <- scan.txt.SpectraSuite(files="data/*.txt")
}

