% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid_angle.R
\name{solid_angle}
\alias{solid_angle}
\title{Calculate 'Solid Angle' of the collection apature}
\usage{
solid_angle(theta_fov)
}
\arguments{
\item{theta_fov}{Angular field of view of the instrument light collection apature in degrees (\eqn{\theta _{fov}}, °)}
}
\value{
A value representing the solid angle of the apature in steradians (Ω, sr)
}
\description{
Calculated as:
\deqn{{\Omega} = \pi \times \sin \left( \frac{\theta_{fov}}{2} \times \frac{\pi}{180} \right)^{2}}{Ω = π × sin(θ~fov~/2 × π/180)^2^}
}
\examples{
solid_angle(20)
}

