% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_spread.R
\name{wavelength_spread}
\alias{wavelength_spread}
\title{Calculate wavelength spread}
\usage{
wavelength_spread(wavelengths)
}
\arguments{
\item{wavelengths}{A vector of wavelengths that represent the center point
of each band}
}
\value{
A vector (same length as \code{wavelengths}) of half-power bandwidths
(half-width of each band) in the units of the given \code{wavelengths} vector
}
\description{
Given a vector of wavelengths that represent the center point
of each band calculate the ‘wavelength spread’ or ‘half-power bandwidths’
for each band (dλ, nm) i.e., the half-width of each band.
  
Calculated as:
\deqn{ d_{\lambda} = \frac{\lambda_{i} - \lambda_{i+1}}{2} }
}
\examples{
wavelength_spread(1:10)
}

