% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_spectra_type.R
\name{assign.type}
\alias{assign.type}
\title{Assign spectra type using a logfile}
\usage{
assign.type(dn, logfile = "fieldlog.csv")
}
\arguments{
\item{dn}{A \pkg{hyperSpec} object}

\item{logfile}{Path to the corresponding logfile for the object}
}
\value{
A \pkg{hyperSpec} object including with sample identification [id] and indication if a reference measurement [is.REF]
}
\description{
Given a log file with columns sample identification, path to reference and sample measurements [id, R, S] this function creates a logical vector [is.REF] with the \pkg{hyperSpec}
}
\details{
The format of the logfile must be followed (.CSV). FIXME make more flexible
}
\examples{
# set path to data files
file.path <- system.file("extdata", package = "FASTSpectra")

# parse spectra into hyperSpec object
dn <- read.txt.OceanOptics(files=paste0(file.path,"/*.txt"))

# assign measurement id and type using fieldlog file
logfile <- paste0(file.path,"/fieldlog.csv")
dn <- assign.type(dn, logfile=logfile)

# split into sample and reference
type <- slot(dn,"data")[["type"]]
dn <- split(x=dn, f=type)
summary(dn)
}

