% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FASTSpectra.R
\docType{package}
\name{FASTspectra}
\alias{FASTspectra}
\alias{FASTspectra-package}
\title{Parse and process Ocean Optics spectra files.}
\description{
FASTSpectra allows you to parse multiple Ocean Optics spectra
, include the instrument generated metadata into a \pkg{hyperSpec} object.
So far only "text with header" format is supported.
}
\details{
The main functions you're likely to need from \pkg{FASTSpectra} are
\code{\link{read.txt.OceanOptics}}, \code{\link{rad.corr}}
and \code{\link{calc.reflectance}}.
}

