% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.calibration.R
\name{import.calibration}
\alias{import.calibration}
\title{Import an instrument or reference panel's calibration data}
\usage{
import.calibration(files = "*.IrradCal", label = list(spc =
  "C_irrad (uJ/count)"))
}
\arguments{
\item{files}{A specific file path or list of file paths. Default is select all text files in working directory \code{"*.IrradCal"}}

\item{label}{A list of form \code{list(parameter=expression(parameter))}. Default is to label the parameter as "C_irrad (uJ/count)"}

\item{type}{One of \code{c("irradiance", "radiance", "ref_panel")}. Default is irradiance.}
}
\value{
A hyperSpec object including a matrix of spectra, metadata extracted from the spectra headers and file information
}
\description{
Import an instrument or reference panel's calibration data
}
\examples{
setwd("~/Desktop/FASTSpectra") # DELETE ME
out <- import.calibration(files="calibration/*.IrradCal")
}

