% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_radiance.R
\name{rad.corr}
\alias{rad.corr}
\title{Radiometric correction"
Convert instrument values, digital number (DN) counts, to radiometric units using instrument specific metadata and calibration values.}
\usage{
rad.corr(dn, cal.DN2L, is.REF = FALSE, cal.REF)
}
\arguments{
\item{dn}{A \pckg{hyperSpec} object that includes a matrix of hyperspectral digital number values [counts] and a dataframe that includes the integration time of each spectrum}

\item{is.REF}{Logical, default is FALSE, if TRUE \code{cal.REF} must be supplied and radiance is returned relative to a 'lambertian' surface with a reflectance factor of 1 throughout the wavelength range of interest}

\item{cal.REF}{Path to a file that contains the reference panel specific calibration of reflectance [-]}

\item{cal.DNtoL}{Path to a file that contains the instrument specific calibration for conversion from counts to radiance [uJ/s]}
}
\value{
A hyperSpec object including a matrix of spectra, metadata extracted from the spectra headers and file information
}
\description{
Radiometric correction"
Convert instrument values, digital number (DN) counts, to radiometric units using instrument specific metadata and calibration values.
}
\examples{
setwd("~/Desktop/FASTSpectra") # DELETE ME
dn <- scan.txt.SpectraSuite(files="data_for_tst/*.txt")
rad <- rad.corr(dn, is.REF=FALSE, cal.DN2L = "calibration/USB2G14742_08202014_VIS_FIB.IrradCal", cal.REF = "calibration/DF25A-5863_SRT-20-050_Reflectance_2008-12-24.txt")
plot(rad)
ref <- rad.corr(dn, is.REF=TRUE, cal.DN2L = "calibration/USB2G14742_08202014_VIS_FIB.IrradCal", cal.REF = "calibration/DF25A-5863_SRT-20-050_Reflectance_2008-12-24.txt")
plot(rad/ref)
}

