% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_radiance.R
\name{rad.corr}
\alias{rad.corr}
\title{Radiometric correction}
\usage{
rad.corr(dn, cal.DN2RadiantFlux, is.REF = FALSE, cal.RRefPanel, coll.area,
  int.time)
}
\arguments{
\item{dn}{A \pkg{hyperSpec} object that includes a matrix of hyperspectral digital number values [counts] and a dataframe that includes the 'integration time' of each spectrum}

\item{cal.DN2RadiantFlux}{Path to a file that contains the instrument specific calibration for conversion from counts to radiance [uJ/count]}

\item{is.REF}{Logical, default is FALSE, if TRUE \code{cal.RRefPanel} must be supplied and radiance is returned relative to a 'lambertian' surface with a reflectance factor of 1 throughout the wavelength range of interest}

\item{cal.RRefPanel}{Path to a file that contains the reference panel specific calibration of reflectance [-]}

\item{coll.area}{Surface area of the instrument [m2]}

\item{int.time}{Integration time of spectral measurement [s]}
}
\value{
A hyperSpec object including a matrix of spectra, metadata extracted from the spectra headers and file information
}
\description{
Convert instrument values, digital number (DN) counts, to radiometric units using instrument specific metadata and calibration values (instrument response functions).
}
\details{
Radiometry terminology and units from \url{https://en.wikipedia.org/wiki/Template:SI_radiometry_units}. Various unit conversions can be applied: normalised DN per wavelength (counts/s), radiant flux (W), radiance (W/sr m2), irradiance (W/m2). These only make sense if the 'measurement geometry', 'integration time' (s) of a measurement, 'surface area of the instrument light apature' (m), and 'instrument-specific response function' in radiant flux or irradiance units (e.g., J/count) is known and supplied to the function.
}
\examples{
setwd("~/Desktop/FASTSpectra") # DELETE ME
dn <- scan.txt.SpectraSuite(files="data_for_tst/*.txt")
rad <- rad.corr(dn, is.REF=FALSE, cal.DN2RadiantFlux = "calibration/USB2G14742_08202014_VIS_FIB.IrradCal", cal.RRefPanel = "calibration/DF25A-5863_SRT-20-050_Reflectance_2008-12-24.txt")
plot(rad)
ref <- rad.corr(dn, is.REF=TRUE, cal.DN2RadiantFlux = "calibration/USB2G14742_08202014_VIS_FIB.IrradCal", cal.RRefPanel = "calibration/DF25A-5863_SRT-20-050_Reflectance_2008-12-24.txt")
plot(rad/ref)
}

